refactor: moving datastore migrate resource to its module
